Begin incorporating error handling/reporting into the parser

+++ src/commonMain/kotlin/org/kson/parser/Parser.kt
@@ -1,6 +1,7 @@
 package org.kson.parser
 
 import org.kson.ast.*
+import org.kson.parser.messages.Message
 
 /**
  * Defines the Kson parser, implemented as a recursive descent parser which directly implements
@@ -8,7 +9,7 @@ import org.kson.ast.*
  *
  * ```
  * kson -> (objectInternals | value) EOF ;
- * objectInternals -> ( keyword value )* ;
+ * objectInternals -> ( keyword value ","? )* ;
  * value -> objectDefinition
  *        | list
  *        | literal
@@ -26,25 +27,37 @@ import org.kson.ast.*
  * See [section 6.2 here](https://craftinginterpreters.com/parsing-expressions.html#recursive-descent-parsing)
  * for guidance on the implementation approach.
  */
-class Parser(tokens: List<Token>) {
+class Parser(tokens: List<Token>, private val messageSink: MessageSink) {
     private val tokenScanner = TokenScanner(tokens)
 
     /**
      * kson -> (objectInternals | value) EOF ;
      */
-    fun parse(): KsonRoot {
+    fun parse(): KsonRoot? {
         val objectInternals = objectInternals()
         if (objectInternals != null) {
-            return KsonRoot(ObjectDefinitionNode(internalsNode = objectInternals))
+            return if (hasUnexpectedTrailingContent()) {
+                null
+            } else {
+                KsonRoot(ObjectDefinitionNode(internalsNode = objectInternals))
+            }
         }
 
-        return KsonRoot(value() ?: TODO("make this a user-friendly parse error"))
-        // parser todo validate we're at EOF to ensure we've parsed everything
+        val value = value()
+        if (value != null) {
+            return if (hasUnexpectedTrailingContent()) {
+                null
+            } else {
+                KsonRoot(value)
+            }
+        }
+
+        // unable to parse
+        return null
     }
 
     /**
-     * objectInternals -> ( keyword value )* ;
-     * parser todo need to support an optional comma here
+     * objectInternals -> ( keyword value ","? )* ;
      */
     private fun objectInternals(): ObjectInternalsNode? {
         val properties = ArrayList<PropertyNode>()
@@ -53,6 +66,11 @@ class Parser(tokens: List<Token>) {
         while (true) {
             val keywordNode = keyword() ?: break
             val valueNode = value() ?: TODO("make this a user-friendly parse error")
+            if (tokenScanner.peek() == TokenType.COMMA) {
+                // drop the optional COMMA
+                tokenScanner.drop()
+            }
+
             properties.add(
                 PropertyNode(
                     keywordNode,
@@ -93,10 +111,18 @@ class Parser(tokens: List<Token>) {
                 ""
             }
 
-            // parser todo syntax error if no brace/malformed object
-            // drop the BRACE_R
-            tokenScanner.drop()
+            val objectStartLocation = tokenScanner.drop()
             val objectInternals = objectInternals()
+            if (tokenScanner.peek() == TokenType.BRACE_R) {
+                // drop the closing brace
+                tokenScanner.drop()
+            } else {
+                messageSink.error(
+                    Location.merge(objectStartLocation, tokenScanner.currentLocation()),
+                    Message.OBJECT_NO_CLOSE
+                )
+                return null
+            }
             return ObjectDefinitionNode(objectName, objectInternals)
         } else {
             // not an objectDefinition
@@ -110,13 +136,12 @@ class Parser(tokens: List<Token>) {
     private fun list(): ListNode? {
         if (tokenScanner.peek() == TokenType.BRACKET_L) {
             // drop the BRACKET_L
-            tokenScanner.drop()
+            val listStartLocation = tokenScanner.drop()
 
             val values = ArrayList<ValueNode>()
             while (tokenScanner.peek() != TokenType.BRACKET_R) {
                 val value = value()
-                    ?:
-                    if (tokenScanner.peek() == TokenType.COMMA) {
+                    ?: if (tokenScanner.peek() == TokenType.COMMA) {
                         // drop the COMMA
                         tokenScanner.drop()
                         continue
@@ -127,9 +152,16 @@ class Parser(tokens: List<Token>) {
                 values.add(value)
             }
 
-            // parser todo syntax error if no bracket/malformed list
-            // drop the BRACKET_R
-            tokenScanner.drop()
+            if (tokenScanner.peek() == TokenType.BRACKET_R) {
+                // drop the BRACKET_R
+                tokenScanner.drop()
+            } else {
+                messageSink.error(
+                    Location.merge(listStartLocation, tokenScanner.currentLocation()),
+                    Message.LIST_NO_CLOSE
+                )
+                return null
+            }
             return ListNode(values)
         } else {
             // not a list
@@ -218,6 +250,30 @@ class Parser(tokens: List<Token>) {
             return null
         }
     }
+
+    /**
+     * Returns true if there is still un-parsed content in [tokenScanner], logging a message
+     * to [messageSink] if it finds unexpected content.  Should only be called to validate the state
+     * of [tokenScanner] after a successful parse
+     */
+    private fun hasUnexpectedTrailingContent(): Boolean {
+        if (tokenScanner.peek() == TokenType.EOF) {
+            // all good: every pre-EOF token has been consumed
+            return false
+        } else {
+            // mark the unexpected content
+            val firstBadTokenLocation = tokenScanner.currentLocation()
+            var lastBadTokenLocation = tokenScanner.drop()
+            while (tokenScanner.peek() != TokenType.EOF) {
+                lastBadTokenLocation = tokenScanner.drop()
+            }
+            messageSink.error(
+                Location.merge(firstBadTokenLocation, lastBadTokenLocation),
+                Message.EOF_NOT_REACHED
+            )
+            return true
+        }
+    }
 }
 
 /**
@@ -256,7 +312,7 @@ private class TokenScanner(private val source: List<Token>) {
      * sequence of tokens.  Note that these [Location]s are pure passthroughs to the
      * [Location]s of the underlying tokens.
      */
-    private fun currentLocation(): Location {
+    fun currentLocation(): Location {
         val startTokenLocation = source[selectionStartOffset].lexeme.location
         val endTokenLocation = source[selectionEndOffset].lexeme.location
         return Location(
